First attempt at a retrying mechanism
We're getting 404s quite frequently from the GitHub API when
we request data from the API for pull request data that appears
to take a while to propagate through the GitHub system.

These errors are usually raised and retried by Sidekiq, and the second
or third attempt is eventually successful. But it causes a lot of noise
in our logs and the errors are too generic to catch/ignore.

This adds a GH wrapper class that will make a specified number
of retries before raising the 404.